+++ gcc/libgdiagnostics.h
@@ -734,6 +734,37 @@ extern diagnostic_file *
 diagnostic_physical_location_get_file (const diagnostic_physical_location *physical_loc)
   LIBGDIAGNOSTICS_PARAM_CAN_BE_NULL(0);
 
+/* Attempt to parse SPEC as if an argument to GCC's
+   -fdiagnostics-add-output=OUTPUT-SPEC.
+   If successful, add an output sink to AFFECTED_MGR and return zero.
+   Otherwise, emit a diagnostic to CONTROL_MGR and return non-zero.
+   Added in LIBGDIAGNOSTICS_ABI_2.  */
+#define LIBDIAGNOSTICS_HAVE_diagnostic_manager_add_sink_from_spec
+
+extern int
+diagnostic_manager_add_sink_from_spec (diagnostic_manager *affected_mgr,
+				       const char *option_name,
+				       const char *spec,
+				       diagnostic_manager *control_mgr)
+  LIBGDIAGNOSTICS_PARAM_MUST_BE_NON_NULL (1)
+  LIBGDIAGNOSTICS_PARAM_MUST_BE_NON_NULL (2)
+  LIBGDIAGNOSTICS_PARAM_MUST_BE_NON_NULL (3)
+  LIBGDIAGNOSTICS_PARAM_MUST_BE_NON_NULL (4);
+
+
+/* Set the main input file of MGR to be FILE.
+   This affects the <title> of generated HTML and
+   the "role" of the artifact in SARIF output (SARIF v2.1.0
+   section 3.24.6).
+   Added in LIBGDIAGNOSTICS_ABI_2.  */
+#define LIBDIAGNOSTICS_HAVE_diagnostic_manager_set_analysis_target
+
+extern void
+diagnostic_manager_set_analysis_target (diagnostic_manager *mgr,
+					const diagnostic_file *file)
+  LIBGDIAGNOSTICS_PARAM_MUST_BE_NON_NULL (1)
+  LIBGDIAGNOSTICS_PARAM_MUST_BE_NON_NULL (2);
+
 /* DEFERRED:
    - thread-safety
    - plural forms
